+++ OneEngine/Game/Rendering/TextRenderer.cs
@@ -3,7 +3,7 @@
     public sealed class TextRenderer : Renderer
     {
         public string Text { get; set; } = "NewText";
-        public int FontSize { get; set; } = 10;
+        public float FontSize { get; set; } = 10;
 
         protected override void OnGraphicsUpdate(IGraphics graphics, Camera camera)
         {
